DEVNULL: * Melhorias e correcoes no suporte a strings (literais) na VM * Correcoes no suporte ao tipo real * Na GptVM, as areas de dados globais e locais foram unificadas

+++ gpt2/gptvm/src/CBytecode.cpp
@@ -15,7 +15,7 @@ CBytecode::CBytecode()
 
 char CBytecode::fetchByte()
 {
-   if (_IP >= size()) {
+   if ((size_t)_IP >= size()) {
       std::cout << "Erro limite em fetchByte()" << std::endl;
       abort();
    }
